backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
click>=8.0.0,<9.0.0
pydantic[email]==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
email-validator==2.1.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
redis==5.2.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0

# Email
sendgrid==6.11.0

# Payments
stripe==11.4.1

# AI/ML (for future CAP 1 & 2)
openai==1.58.1
pgvector==0.3.6
numpy==2.4.1

# Resume Parsing
PyPDF2==3.0.1
python-docx==1.1.2

# Graph Database
neo4j==5.27.0

# Utilities
python-dotenv==1.0.1
httpx==0.28.1
slowapi==0.1.9

# Development
pytest==8.3.4
pytest-asyncio==0.25.2
black==24.10.0
ruff==0.8.4
